Refactor: FolderPage 수정
