--- src/styles/LinkCard.css
@@ -1,80 +1,0 @@
-.img-group {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 20rem;
-  overflow: hidden;
-}
-
-.img-card {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.img-card.no-img {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #dddfff;
-}
-
-.img-card.no-img::after {
-  content: '';
-  display: block;
-  width: 13.3rem;
-  height: 2.4rem;
-  background: url('../assets//linkbrary-logo.svg') no-repeat center/cover;
-  opacity: 0.2;
-}
-
-.btn-star {
-  position: absolute;
-  top: 1.5rem;
-  right: 1.5rem;
-}
-
-.card-info-group {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1.5rem 2rem;
-}
-
-.time-info-group {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.elapsed-time {
-  line-height: 1.7rem;
-  font-size: 1.3rem;
-  font-weight: 400;
-  color: #666;
-}
-
-.more {
-  display: flex;
-}
-
-.desc {
-  display: -webkit-box;
-  height: 4.9rem;
-  font-size: 1.6rem;
-  font-weight: 400;
-  line-height: 2.4rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 2;
-}
-
-.date {
-  line-height: 1.9rem;
-  font-size: 14px;
-  font-weight: 400;
-  color: #333;
-}
